lib/resources/requirements.txt
boto3
fpdf
Pillow
python-resize-image
requests